AO3-6918 Refactor routing
