+++ app/views/admin/preferences/totp_setup.erb
@@ -16,7 +16,7 @@
 <fieldset>
   <h3 class="heading"><%= t(".step_2.heading") %></h3>
   <p><%= t(".step_2.instructions") %></p>
-  <%= form_for(current_admin, url: totp_setup_form_admin_path, method: :post) do |form| %>
+  <%= form_for(current_admin, url: totp_setup_form_admin_preferences_path, method: :post) do |form| %>
     <dl>
       <dt><%= form.label :password, t(".step_2.account_password") %></dt>
       <dd><%= form.password_field :password %></dd>
