+++ app/views/users/sessions/_greeting.html.erb
@@ -53,7 +53,7 @@
         <li class="dropdown">
           <%= link_to t(".nav.greeting", current_user: current_admin.login), admins_path %>
           <ul class="menu">
-            <li><%= link_to t(".nav.account"), preferences_admin_path(current_admin) %></li>
+            <li><%= link_to t(".nav.account"), admin_preferences_path(current_admin) %></li>
           </ul>
         </li>
         <li><%= link_to t(".nav.log_out"), destroy_admin_session_path, method: :delete %></li>
